Sliver app bar in maintenance page
