+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Sliver app bar in entity definitions
 - Sliver app bar in speech settings
 - Sliver app bar in config flags page
+- Sliver app bar in maintenance page
 
 ## [0.8.355] - 2023-05-19
 ### Added:
