+++ lib/beamer/locations/settings_location.dart
@@ -1,12 +1,14 @@
 import 'package:beamer/beamer.dart';
 import 'package:flutter/material.dart';
 import 'package:lotti/pages/journal/entry_details_page.dart';
-import 'package:lotti/pages/settings/about_page.dart';
+import 'package:lotti/pages/settings/advanced/about_page.dart';
+import 'package:lotti/pages/settings/advanced/conflicts_page.dart';
+import 'package:lotti/pages/settings/advanced/logging_page.dart';
+import 'package:lotti/pages/settings/advanced/maintenance_page.dart';
 import 'package:lotti/pages/settings/advanced_settings_page.dart';
 import 'package:lotti/pages/settings/categories/categories_page.dart';
 import 'package:lotti/pages/settings/categories/category_create_page.dart';
 import 'package:lotti/pages/settings/categories/category_details_page.dart';
-import 'package:lotti/pages/settings/conflicts_page.dart';
 import 'package:lotti/pages/settings/dashboards/create_dashboard_page.dart';
 import 'package:lotti/pages/settings/dashboards/dashboard_definition_page.dart';
 import 'package:lotti/pages/settings/dashboards/dashboards_page.dart';
@@ -15,8 +17,6 @@ import 'package:lotti/pages/settings/habits/habit_create_page.dart';
 import 'package:lotti/pages/settings/habits/habit_details_page.dart';
 import 'package:lotti/pages/settings/habits/habits_page.dart';
 import 'package:lotti/pages/settings/health_import_page.dart';
-import 'package:lotti/pages/settings/logging_page.dart';
-import 'package:lotti/pages/settings/maintenance_page.dart';
 import 'package:lotti/pages/settings/measurables/measurable_create_page.dart';
 import 'package:lotti/pages/settings/measurables/measurable_details_page.dart';
 import 'package:lotti/pages/settings/measurables/measurables_page.dart';
