+++ lib/pages/settings/advanced/about_page.dart
