+++ lib/pages/settings/advanced/conflicts_page.dart
