+++ lib/pages/settings/advanced/logging_page.dart
