+++ test/pages/settings/maintenance_page_test.dart
@@ -3,7 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/database/maintenance.dart';
 import 'package:lotti/get_it.dart';
-import 'package:lotti/pages/settings/maintenance_page.dart';
+import 'package:lotti/pages/settings/advanced/maintenance_page.dart';
 import 'package:lotti/themes/themes_service.dart';
 import 'package:mocktail/mocktail.dart';
